+++ library/foundation/src/vscf_key_recipient_info_list.c
@@ -271,6 +271,26 @@ vscf_key_recipient_info_list_add(
 }
 
 //
+//  Remove current node.
+//
+VSCF_PRIVATE void
+vscf_key_recipient_info_list_remove_self(vscf_key_recipient_info_list_t *self) {
+
+    VSCF_ASSERT_PTR(self);
+
+    vscf_key_recipient_info_destroy(&self->item);
+    if (self->next) {
+        vscf_key_recipient_info_list_t *next = self->next;
+        self->item = next->item;
+        self->next = next->next;
+        next->next = NULL; //  prevent chain destruction
+        next->item = NULL;
+        next->prev = NULL;
+        vscf_key_recipient_info_list_destroy(&next);
+    }
+}
+
+//
 //  Return true if given list has item.
 //
 VSCF_PUBLIC bool
@@ -309,6 +329,17 @@ vscf_key_recipient_info_list_has_next(const vscf_key_recipient_info_list_t *self
 //
 VSCF_PUBLIC vscf_key_recipient_info_list_t *
 vscf_key_recipient_info_list_next(const vscf_key_recipient_info_list_t *self) {
+
+    VSCF_ASSERT_PTR(self);
+
+    return self->next;
+}
+
+//
+//  Return next list node if exists, or NULL otherwise.
+//
+VSCF_PRIVATE vscf_key_recipient_info_list_t *
+vscf_key_recipient_info_list_next_modifiable(vscf_key_recipient_info_list_t *self) {
 
     VSCF_ASSERT_PTR(self);