+++ wrappers/python/virgil_crypto_lib/phe/common.py
@@ -58,3 +58,5 @@ class Common(object):
     PHE_MAX_ENCRYPT_LEN = 1024 * 1024 - 64
     # Maximum data size to decrypt
     PHE_MAX_DECRYPT_LEN = 1024 * 1024
+    # Maximum data to authenticate
+    PHE_MAX_AUTH_LEN = 1024