+++ wrappers/wasm/foundation/src/asmjs.js
@@ -66,6 +66,7 @@ const initBrainkeyServer = require('./BrainkeyServer');
 const initGroupSessionMessage = require('./GroupSessionMessage');
 const initGroupSessionTicket = require('./GroupSessionTicket');
 const initGroupSession = require('./GroupSession');
+const initMessageInfoEditor = require('./MessageInfoEditor');
 const initSha224 = require('./Sha224');
 const initSha256 = require('./Sha256');
 const initSha384 = require('./Sha384');
@@ -142,6 +143,7 @@ const initProject = () => {
     modules.GroupSessionMessage = initGroupSessionMessage(foundationModule, modules);
     modules.GroupSessionTicket = initGroupSessionTicket(foundationModule, modules);
     modules.GroupSession = initGroupSession(foundationModule, modules);
+    modules.MessageInfoEditor = initMessageInfoEditor(foundationModule, modules);
     modules.Sha224 = initSha224(foundationModule, modules);
     modules.Sha256 = initSha256(foundationModule, modules);
     modules.Sha384 = initSha384(foundationModule, modules);